Creates Offsets module and moves methods/tests
